move client impl into separate file: metering_client_http + make test run
